test(e2e): refactor Playwright test setup and helpers
Extract common setup logic into a dedicated helper function `setupTestEnvironment`.
Introduce new helper files for chat interactions and consolidate exports.
This improves test readability and reduces duplication in E2E tests.
